refactor(checker): split tuple-chain indexed access helpers
